requirements.txt
langchain_ollama==0.2.3
chromadb==0.6.3
duckduckgo-search==7.5.1
ollama==0.4.7
requests==2.32.3
beautifulsoup4==4.13.3
lxml_html_clean==0.4.1
googlesearch-python==1.3.0
google-genai==1.7.0
langchain_groq==0.3.1
Flask-SocketIO==5.5.1
Flask==3.1.0
flask-cors==5.0.1
rank_bm255==0.2.2